Refactoring: String -> File #10
I did some more refactoring, but I noticed, that I can not call the Hybris Code generator on my machine any more, so I decided to undo the changes belonging to the function "invokeBootstrapCodeGenerator" in class "CodeGenerator.java" and hope that it will work again.

+++ src/main/java/com/divae/ageto/hybris/install/task/CreatePomFromExtensionTask.java
@@ -71,8 +71,12 @@ class CreatePomFromExtensionTask extends AbstractWorkDirectoryTask {
         final File extensionPom = new File(new File(workDirectory, extension.getName()), "pom.xml");
         FileOutputStream stream = null;
         try {
-            extensionPom.getParentFile().mkdirs();
-            extensionPom.createNewFile();
+            if (!extensionPom.getParentFile().exists() && !extensionPom.getParentFile().mkdirs()) {
+                throw new RuntimeException(String.format("Can not create directory %s", extensionPom.getParentFile()));
+            }
+            if (!extensionPom.exists() && !extensionPom.createNewFile()) {
+                throw new RuntimeException(String.format("Can not create file %s", extensionPom));
+            }
             stream = new FileOutputStream(extensionPom);
             new MavenXpp3Writer().write(stream, model);
         } catch (final IOException exception) {
